requirements.txt
Flask==1.1.1
gunicorn==19.6.0
line-bot-sdk==1.3.0